APIv2: Add support for configuration_permissions
