+++ unittests/test_apiv2_methods.py
@@ -16,7 +16,7 @@ class ApiEndpointMethods(DojoTestCase):
     def test_is_defined(self):
 
         for reg, _, _ in sorted(self.registry):
-            if reg in ['import-scan', 'reimport-scan', 'notes', 'system_settings', 'roles', 'import-languages', 'endpoint_meta_import', 'test_types']:
+            if reg in ['import-scan', 'reimport-scan', 'notes', 'system_settings', 'roles', 'import-languages', 'endpoint_meta_import', 'test_types', 'configuration_permissions']:
                 continue
 
             for method in ['get', 'post']:
